src/main/java/com/mycompany/igorfood/domain/service/CadastroCozinhaService.java
package com.mycompany.igorfood.domain.service;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.dao.DataIntegrityViolationException;
import org.springframework.dao.EmptyResultDataAccessException;
import org.springframework.stereotype.Service;

import com.mycompany.igorfood.domain.exception.EntidadeEmUsoException;
import com.mycompany.igorfood.domain.exception.EntidadeNaoEncontradaException;
import com.mycompany.igorfood.domain.model.Cozinha;
import com.mycompany.igorfood.domain.repository.CozinhaRepository;

@Service
public class CadastroCozinhaService {
	
	@Autowired
	private CozinhaRepository cozinhaRepository;
	
	public Cozinha salvar(Cozinha cozinha) {
		return cozinhaRepository.salvar(cozinha);
	}
	
	public void excluir(Long cozinhaId) {
		try {
			cozinhaRepository.remover(cozinhaId);			
		} catch (EmptyResultDataAccessException e) {
			throw new EntidadeNaoEncontradaException(
				String.format("Não existe cadastro de cozinha com código %d", cozinhaId));
			
		} catch (DataIntegrityViolationException e) {
			throw new EntidadeEmUsoException(
				String.format("Cozinha de código %d não pode ser removida. Está em uso.", cozinhaId));
		}
		
	}
	
	public String teste() {
		return "Teste";
	}

}
